pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "neewer"
version = "0.1.0"
description = "BLE control interface for Neewer LED lights"
requires-python = ">=3.10"
dependencies = [
    "bleak>=0.21.0",
    "sacn>=1.9.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
tui = ["textual>=0.80.0"]
audio = ["numpy>=1.24.0", "sounddevice>=0.4.6"]
dev = ["pytest>=7.0"]

[project.scripts]
neewer = "neewer.protocol:main_sync"
neewer-sacn = "neewer.sacn:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
